fix(core): 修复流量数据显示类型错误问题
- 将版本号从 2.3.0-preview3 更新为 2.3.0-preview4
- 将 MEFrpVersion 从可变属性改为常量定义
- 移除未使用的 LogService 静态字段
- 将流量数据类型从 int 数组改为 ulong 数组以支持更大数值
- 更新流量状态控件中的图表系列数据类型匹配
- 确保流量显示图表的数据类型一致性

+++ MEFrpLauncherX/Controls/TrafficStatusControl.axaml.cs
@@ -158,16 +158,16 @@ public class TrafficStatusControlViewModel : ViewModelBase
             switch (type)
             {
                 case 0:
-                    newSeries.Add(new LineSeries<int>()
+                    newSeries.Add(new LineSeries<ulong>()
                     {
-                        Values = new AvaloniaList<int>(trafficIn),
+                        Values = new AvaloniaList<ulong>(trafficIn),
                         YToolTipLabelFormatter = x => ProcessFileSize(x.Model),
                         Name = "入站流量"
                     }); break;
                 case 1:
-                    newSeries.Add(new ColumnSeries<int>()
+                    newSeries.Add(new ColumnSeries<ulong>()
                     {
-                        Values = new AvaloniaList<int>(trafficIn),
+                        Values = new AvaloniaList<ulong>(trafficIn),
                         YToolTipLabelFormatter = x => ProcessFileSize(x.Model),
                         Name = "入站流量"
                     }); break;
@@ -179,16 +179,16 @@ public class TrafficStatusControlViewModel : ViewModelBase
             switch (type)
             {
                 case 0:
-                    newSeries.Add(new LineSeries<int>()
+                    newSeries.Add(new LineSeries<ulong>()
                     {
-                        Values = new AvaloniaList<int>(trafficOut),
+                        Values = new AvaloniaList<ulong>(trafficOut),
                         YToolTipLabelFormatter = x => ProcessFileSize(x.Model),
                         Name = "出站流量"
                     }); break;
                 case 1:
-                    newSeries.Add(new ColumnSeries<int>()
+                    newSeries.Add(new ColumnSeries<ulong>()
                     {
-                        Values = new AvaloniaList<int>(trafficOut),
+                        Values = new AvaloniaList<ulong>(trafficOut),
                         YToolTipLabelFormatter = x => ProcessFileSize(x.Model),
                         Name = "出站流量"
                     }); break;
@@ -200,16 +200,16 @@ public class TrafficStatusControlViewModel : ViewModelBase
             switch (type)
             {
                 case 0:
-                    newSeries.Add(new LineSeries<int>()
+                    newSeries.Add(new LineSeries<ulong>()
                     {
-                        Values = new AvaloniaList<int>(totalTraffic),
+                        Values = new AvaloniaList<ulong>(totalTraffic),
                         YToolTipLabelFormatter = x => ProcessFileSize(x.Model),
                         Name = "总流量"
                     }); break;
                 case 1:
-                    newSeries.Add(new ColumnSeries<int>()
+                    newSeries.Add(new ColumnSeries<ulong>()
                     {
-                        Values = new AvaloniaList<int>(totalTraffic),
+                        Values = new AvaloniaList<ulong>(totalTraffic),
                         YToolTipLabelFormatter = x => ProcessFileSize(x.Model),
                         Name = "总流量"
                     }); break;
